Refactor particles
